Move GetDNSIP to a "higher level" pkg.
